requirements.txt
numpy
scipy
pywavelets
matplotlib
Pillow
bitarray
requests
python-dotenv